# src/BotNexus/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/admin/fission/config": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取系统的裂变策略配置，如积分规则、开关状态等",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Fission"
                ],
                "summary": "获取裂变全局配置",
                "responses": {
                    "200": {
                        "description": "裂变配置",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "更新系统的裂变策略配置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Fission"
                ],
                "summary": "更新裂变全局配置",
                "parameters": [
                    {
                        "description": "配置信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.FissionConfigGORM"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "$ref": "#/definitions/utils.JSONResponse"
                        }
                    }
                }
            }
        },
        "/api/admin/fission/invitations": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取系统的最近 100 条用户邀请记录",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Fission"
                ],
                "summary": "获取邀请记录",
                "responses": {
                    "200": {
                        "description": "邀请记录列表",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/admin/fission/leaderboard": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取系统邀请人数排名前 20 的用户记录",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Fission"
                ],
                "summary": "获取裂变排行榜",
                "responses": {
                    "200": {
                        "description": "排行榜数据",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/admin/fission/stats": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取裂变总数、积分总额、活跃邀请人及最近 7 天趋势",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Fission"
                ],
                "summary": "获取裂变统计",
                "responses": {
                    "200": {
                        "description": "统计数据",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/admin/fission/tasks": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取所有定义的裂变任务列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Fission"
                ],
                "summary": "获取裂变任务",
                "responses": {
                    "200": {
                        "description": "任务列表",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "新增或更新裂变任务配置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Fission"
                ],
                "summary": "保存裂变任务",
                "parameters": [
                    {
                        "description": "任务信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.FissionTaskGORM"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "保存成功",
                        "schema": {
                            "$ref": "#/definitions/utils.JSONResponse"
                        }
                    }
                }
            }
        },
        "/api/admin/fission/tasks/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据 ID 删除指定的裂变任务配置",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Fission"
                ],
                "summary": "删除裂变任务",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "任务 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "$ref": "#/definitions/utils.JSONResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.FissionConfigGORM": {
            "type": "object",
            "properties": {
                "anti_fraud_enabled": {
                    "type": "boolean"
                },
                "enabled": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "invite_code_template": {
                    "type": "string"
                },
                "invite_reward_duration": {
                    "description": "小时",
                    "type": "integer"
                },
                "invite_reward_points": {
                    "type": "integer"
                },
                "max_daily_invites": {
                    "type": "integer"
                },
                "min_level_required": {
                    "type": "integer"
                },
                "new_user_reward_points": {
                    "type": "integer"
                },
                "rules": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "welcome_message": {
                    "type": "string"
                }
            }
        },
        "models.FissionTaskGORM": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "reward_duration": {
                    "type": "integer"
                },
                "reward_points": {
                    "type": "integer"
                },
                "status": {
                    "description": "active, inactive",
                    "type": "string"
                },
                "target_count": {
                    "type": "integer"
                },
                "task_type": {
                    "description": "register, use_bot, join_group",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "utils.JSONResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    }
}